backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
qdrant-client==1.7.0
psycopg2-binary==2.9.9
pydantic==2.5.0
httpx==0.25.1
markdown==3.5.1
beautifulsoup4==4.12.2
tiktoken==0.5.1
python-multipart==0.0.6
